s_plot funtionality
