+++ printhor/src/bin/instrumentation/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod datapoints;
+pub(crate) mod machinery;
+pub(crate) mod gcode;
